Bug recipe's dishes and cuisine partly solved
