+++ src/screens/MenuScreen.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { StyleSheet, Dimensions, View, Text } from "react-native";
 import { useFocusEffect, useIsFocused } from "@react-navigation/native";
 import Constants from "expo-constants";
@@ -9,6 +9,8 @@ import { colorPalette, shadowStyle } from "../constants/ColorPalette";
 import CardStack, { Card } from "react-native-card-stack-swiper";
 import RecipeCard from "../components/RecipeCard";
 import SwipeButtons from "../components/SwipeButtons";
+import RecipeCardStack from "../components/RecipeCardStack"
+import LoadingCardStack from "../components/LoadingCardStack"
 
 // Importing JSON data for development and testing
 import * as recipesJson from "../data/recipes.json";
@@ -24,16 +26,10 @@ const randRecipeUrl = `https://api.spoonacular.com/recipes/random?apiKey=${API_K
 
 export default function MenuScreen() {
     // useDispatch allows us to dispatch Actions to mutate global store variables
-    const dispatch = useDispatch();
     const [randRecipes, setRandRecipes] = React.useState<Recipe[]>([initialState.recipe]);
-    const userState = useSelector<RootState, UserState>((state) => state.userState);
+    const [isCardStackLoading, setIsCardStackLoading] = React.useState<boolean>(false);
     const filtersState = useSelector<RootState, FiltersState>((state) => state.filtersState);
-    const userRecipeListState = useSelector<RootState, UserRecipeListState>((state) => state.userRecipeListState);
 
-    // FOR TEST PURPOSES
-    const [swipedLeftRecipes, setSwipedLeftRecipes] = React.useState<Recipe[]>([]);
-    const [swipedRightRecipes, setSwipedRightRecipes] = React.useState<Recipe[]>([]);
-    let cardStackRef = React.useRef<CardStack | any>();
 
     // Fetch random Recipes from Spoonacular
     async function fetchRandomRecipes() {
@@ -71,12 +67,14 @@ export default function MenuScreen() {
         })
 
         setRandRecipes(filteredRecipes);
+        setIsCardStackLoading(false)
     }
 
     // On focus, fetch/set random Recipes
     useFocusEffect(
         React.useCallback(() => {
             console.log("MenuScreen Focused")
+            fetchRandomRecipes()
         }, [])
     );
 
@@ -84,167 +82,63 @@ export default function MenuScreen() {
     React.useEffect(() => {
         console.log("FROM MENU SCREEN -> dishType:", filtersState.filters.dishType);
         // Fetch new recipes and apply new filters
-        fetchRandomRecipes();
-        console.log("MenuScreen Filters changed")
-        // Re-render the page to update entire cardstack with new stateful values
-      }, [filtersState.filters])
+
+        console.log("MenuScreen Filters changed");
+        // Re-render the page to update entire cardstack with new stateful value
+
+        setIsCardStackLoading(true)
+        changeRandomRecipes();
+    }, [filtersState])
+
+    // Helper function 
+    function shuffleRecipes(array: Recipe[]) {
+        array.sort(() => Math.random() - .5);
+    }
+
+    function changeRandomRecipes() {
+        const newRandomRecipes = recipesJson.recipes.map((rcp) => {
+            return {
+                id: rcp.id,
+                sourceUrl: rcp.sourceUrl,
+                image: rcp.image,
+                imageType: rcp.imageType,
+                title: rcp.title,
+                diets: rcp.diets,
+                cuisines: rcp.cuisines,
+                dishTypes: rcp.dishTypes,
+                vegetarian: rcp.vegetarian,
+                vegan: rcp.vegan,
+                glutenFree: rcp.glutenFree,
+                dairyFree: rcp.dairyFree,
+                veryHealthy: rcp.veryHealthy,
+                cheap: rcp.cheap,
+                veryPopular: rcp.veryPopular,
+                sustainable: rcp.sustainable,
+                aggregateLikes: rcp.aggregateLikes,
+                spoonacularScore: rcp.spoonacularScore,
+                healthScore: rcp.healthScore,
+                pricePerServing: rcp.pricePerServing,
+                readyInMinutes: rcp.readyInMinutes,
+                servings: rcp.servings,
+            }
+        })
+
+        shuffleRecipes(newRandomRecipes);
+        setRandRecipes(newRandomRecipes);
+        setIsCardStackLoading(false);
+    }
 
     console.log("OUTSIDE MENU SCREEN -> dishType:", filtersState.filters.dishType);
     // Listen to when randRecipes get set
-    async function onSwipedLeft(idx: number) {
-        console.log("Swiped left");
-        // const query = `query getUser {
-        //     users {
-        //       id
-        //       username
-        //     }
-        //   }`
-
-        // const user = await axios({
-        //     url: "https://savored-server.herokuapp.com/",
-        //     method: "post",
-        //     data: {
-        //         query: query,
-        //     }
-        // })
-
-        // TODO: store it the database instead
-        console.log("onSwipedLeft -> dishType:", filtersState.filters.dishType);
-
-        const recipeToBeAdded = {
-            id: randRecipes[idx].id,
-            title: randRecipes[idx].title,
-            // Change cuisne and dishType to be comma separated string value
-            cuisine: filtersState.filters.cuisine === "" ? randRecipes[idx].cuisines.toString() : filtersState.filters.cuisine,
-            dishType: filtersState.filters.dishType === "" ? randRecipes[idx].dishTypes.toString() : filtersState.filters.dishType,
-            vegetarian: randRecipes[idx].vegetarian,
-            vegan: randRecipes[idx].vegan,
-            glutenFree: randRecipes[idx].glutenFree,
-            dairyFree: randRecipes[idx].dairyFree,
-            readyInMinutes: randRecipes[idx].readyInMinutes,
-            servings: randRecipes[idx].servings,
-            isSavored: false
-        }
-
-
-        dispatch(addtoUserRecipeList(recipeToBeAdded))
-
-        swipedLeftRecipes.push(randRecipes[idx]);
-        setSwipedLeftRecipes(swipedLeftRecipes);
-    }
-
-    async function onSwipedRight(idx: number) {
-        console.log("Swiped right");
-        // const query = `query getUser {
-        //     users {
-        //       id
-        //       username
-        //     }
-        //   }`
-
-        // const user = await axios({
-        //     url: "https://savored-server.herokuapp.com/",
-        //     method: "post",
-        //     data: {
-        //         query: query,
-        //     }
-        // })
-
-        // TODO: store it the database instead
-        console.log("onSwipedRight -> dishType:", filtersState.filters.dishType);
-
-        const recipeToBeAdded = {
-            id: randRecipes[idx].id,
-            title: randRecipes[idx].title,
-            cuisine: filtersState.filters.cuisine === "" ? randRecipes[idx].cuisines.toString() : filtersState.filters.cuisine,
-            dishType: filtersState.filters.dishType === "" ? randRecipes[idx].dishTypes.toString() : filtersState.filters.dishType,
-            vegetarian: randRecipes[idx].vegetarian,
-            vegan: randRecipes[idx].vegan,
-            glutenFree: randRecipes[idx].glutenFree,
-            dairyFree: randRecipes[idx].dairyFree,
-            readyInMinutes: randRecipes[idx].readyInMinutes,
-            servings: randRecipes[idx].servings,
-            isSavored: true
-        }
-
-
-        dispatch(addtoUserRecipeList(recipeToBeAdded))
-
-        swipedRightRecipes.push(randRecipes[idx])
-        setSwipedRightRecipes(swipedRightRecipes);
-    }
-
-    function handleOnPressLeft() {
-        cardStackRef.current.swipeLeft();
-    }
-
-    function handleOnPressRight() {
-        cardStackRef.current.swipeRight();
-    }
-
-
     // TODO: 
     // - On load/before render make API requests for randomized Recipes (Spoonacular)
     // - Apply filters
     // - Compare against User"s viewed Recipes list if User is logged in
     // - Apply score and sorting if smart filter is turned on
     return (
-        <View style={styles.container}>
-            <View style={styles.subContainer}>
-                {/* {cardsLoading? (
-                    <LoadingCard />
-                ) : (
-                    <RecipeCardStack filtersState={filtersState} />
-                )} */}
-                <CardStack 
-                style={styles.cardStack} 
-                ref={(cardStack: CardStack) => { cardStackRef.current = cardStack }} 
-                renderNoMoreCards={() => { return <Text>No More Recipes</Text> }} 
-                disableBottomSwipe 
-                disableTopSwipe>
-                    {randRecipes.map((rcp: Recipe, idx: number) => {
-                        return (
-                            <Card key={rcp.id} onSwipedLeft={() => { onSwipedLeft(idx) }} onSwipedRight={() => { onSwipedRight(idx) }}>
-                                <RecipeCard rcp={rcp} id={rcp.id} />
-                            </Card>
-                        )
-                    })}
-                </CardStack>
-            </View>
-            <SwipeButtons handleOnPressLeft={handleOnPressLeft} handleOnPressRight={handleOnPressRight} />
-        </View>
+
+        isCardStackLoading ?
+            <LoadingCardStack /> :
+            <RecipeCardStack randRecipes={randRecipes} filtersState={filtersState} />
     )
 }
-
-
-const styles = StyleSheet.create({
-
-    container: {
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-        backgroundColor: colorPalette.background
-    },
-
-    subContainer: {
-        justifyContent: "center",
-        alignItems: "center",
-        width: _screen.width * 0.9,
-        height: _screen.height * 0.6,
-        borderRadius: 30,
-        backgroundColor: colorPalette.primary,
-        ...shadowStyle
-    },
-
-    recipeTextTest: {
-        justifyContent: "center",
-        alignItems: "center",
-        textAlign: "center",
-        margin: 8
-    },
-
-    cardStack: {
-        justifyContent: "center",
-        alignItems: "center"
-    }
-})
